all: add some todos and fix small issues

+++ pkg/connector/store/upgrades/00-latest.sql
@@ -1,22 +1,22 @@
 -- v0 -> v3: Latest revision
 
 CREATE TABLE telegram_session (
-    user_id      INTEGER PRIMARY KEY,
+    user_id      BIGINT PRIMARY KEY,
     session_data BYTEA NOT NULL
 );
 
 CREATE TABLE telegram_user_state (
-    user_id INTEGER PRIMARY KEY,
-    pts     INTEGER NOT NULL,
-    qts     INTEGER NOT NULL,
-    date    INTEGER NOT NULL,
-    seq     INTEGER NOT NULL
+    user_id BIGINT PRIMARY KEY,
+    pts     BIGINT NOT NULL,
+    qts     BIGINT NOT NULL,
+    date    BIGINT NOT NULL,
+    seq     BIGINT NOT NULL
 );
 
 CREATE TABLE telegram_channel_state (
-    user_id    INTEGER,
-    channel_id INTEGER,
-    pts        INTEGER NOT NULL,
+    user_id    BIGINT,
+    channel_id BIGINT,
+    pts        BIGINT NOT NULL,
 
     PRIMARY KEY (user_id, channel_id)
 );
@@ -24,18 +24,18 @@ CREATE TABLE telegram_channel_state (
 CREATE INDEX idx_telegram_channel_state_user_id ON telegram_channel_state (user_id);
 
 CREATE TABLE telegram_channel_access_hashes (
-    user_id     INTEGER,
-    channel_id  INTEGER,
-    access_hash INTEGER NOT NULL,
+    user_id     BIGINT,
+    channel_id  BIGINT,
+    access_hash BIGINT NOT NULL,
 
     PRIMARY KEY (user_id, channel_id)
 );
 
 CREATE TABLE telegram_user_metadata (
-    receiver_id INTEGER,
-    user_id     INTEGER,
+    receiver_id BIGINT,
+    user_id     BIGINT,
 
-    access_hash INTEGER NOT NULL,
+    access_hash BIGINT NOT NULL,
     username    TEXT,
 
     PRIMARY KEY (receiver_id, user_id)
@@ -48,4 +48,5 @@ CREATE TABLE telegram_file (
     size      BIGINT
 );
 
+-- TODO this will be unnecessary once the queries switch to reading telegram_user_metadata
 CREATE INDEX idx_ghost_username ON ghost ((metadata->>'username'));
